do as advised to

+++ data_test.go
@@ -21,28 +21,28 @@ func TestDataSetGet(t *testing.T) {
 
 	for _, want := range []int8{-126, 126} {
 		d.Set(want)
-		if got := d.Get(); got.(int8) != want {
+		if got := d.Get(); got.(int64) != int64(want) {
 			t.Errorf("set %v, got %v", want, got)
 		}
 	}
 
 	for _, want := range []int16{-30000, -15000, 0, 15000, 30000} {
 		d.Set(want)
-		if got := d.Get(); got.(int16) != want {
+		if got := d.Get(); got.(int64) != int64(want) {
 			t.Errorf("set %v, got %v", want, got)
 		}
 	}
 
 	for _, want := range []uint8{0, 128, 253} {
 		d.Set(want)
-		if got := d.Get(); got.(uint8) != want {
+		if got := d.Get(); got.(uint64) != uint64(want) {
 			t.Errorf("set %v, got %v", want, got)
 		}
 	}
 
 	for _, want := range []uint16{0, 15000, 30000, 45000, 60000} {
 		d.Set(want)
-		if got := d.Get(); got.(uint16) != want {
+		if got := d.Get(); got.(uint64) != uint64(want) {
 			t.Errorf("set %v, got %v", want, got)
 		}
 	}
